Move Task to its own module
